// src/api/api.ts
export interface CommandDescription {
  /**
   *
   * @type {Array<CommandOption>}
   * @memberof CommandDescription
   */
  commandOptions?: Array<CommandOption>;
}
// ...
export interface EnrichedItemDTO {
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  type?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  label?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  category?: string;
  /**
   *
   * @type {Array<string>}
   * @memberof EnrichedItemDTO
   */
  tags?: Array<string>;
  /**
   *
   * @type {Array<string>}
   * @memberof EnrichedItemDTO
   */
  groupNames?: Array<string>;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  link?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  state?: string;
  /**
   *
   * @type {string}
   * @memberof EnrichedItemDTO
   */
  transformedState?: string;
  /**
   *
   * @type {StateDescription}
   * @memberof EnrichedItemDTO
   */
  stateDescription?: StateDescription;
  /**
   *
   * @type {CommandDescription}
   * @memberof EnrichedItemDTO
   */
  commandDescription?: CommandDescription;
  /**
   *
   * @type {{ [key: string]: any; }}
   * @memberof EnrichedItemDTO
   */
  metadata?: { [key: string]: any };
  /**
   *
   * @type {boolean}
   * @memberof EnrichedItemDTO
   */
  editable?: boolean;
}
// ...
export interface MappingDTO {
  /**
   *
   * @type {string}
   * @memberof MappingDTO
   */
  command?: string;
  /**
   *
   * @type {string}
   * @memberof MappingDTO
   */
  label?: string;
}

/**
 *
 * @export
 * @interface MetadataDTO
 */
export interface MetadataDTO {
  /**
   *
   * @type {string}
   * @memberof MetadataDTO
   */
  value?: string;
  /**
   *
   * @type {{ [key: string]: any; }}
   * @memberof MetadataDTO
   */
  config?: { [key: string]: any };
}

/**
 *
 * @export
 * @interface PageDTO
 */
export interface PageDTO {
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  title?: string;
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  icon?: string;
  /**
   *
   * @type {string}
   * @memberof PageDTO
   */
  link?: string;
  /**
   *
   * @type {PageDTO}
   * @memberof PageDTO
   */
  parent?: PageDTO;
  /**
   *
   * @type {boolean}
   * @memberof PageDTO
   */
  leaf?: boolean;
  /**
   *
   * @type {boolean}
   * @memberof PageDTO
   */
  timeout?: boolean;
  /**
   *
   * @type {Array<WidgetDTO>}
   * @memberof PageDTO
   */
  widgets?: Array<WidgetDTO>;
}
// ...
export interface StateDescription {
  /**
   *
   * @type {number}
   * @memberof StateDescription
   */
  minimum?: number;
  /**
   *
   * @type {number}
   * @memberof StateDescription
   */
  maximum?: number;
  /**
   *
   * @type {number}
   * @memberof StateDescription
   */
  step?: number;
  /**
   *
   * @type {string}
   * @memberof StateDescription
   */
  pattern?: string;
  /**
   *
   * @type {boolean}
   * @memberof StateDescription
   */
  readOnly?: boolean;
  /**
   *
   * @type {Array<StateOption>}
   * @memberof StateDescription
   */
  options?: Array<StateOption>;
}
// ...
export interface WidgetDTO {
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  widgetId?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  type?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  name?: string;
  /**
   *
   * @type {boolean}
   * @memberof WidgetDTO
   */
  visibility?: boolean;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  label?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  icon?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  labelcolor?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  valuecolor?: string;
  /**
   *
   * @type {Array<MappingDTO>}
   * @memberof WidgetDTO
   */
  mappings?: Array<MappingDTO>;
  /**
   *
   * @type {boolean}
   * @memberof WidgetDTO
   */
  switchSupport?: boolean;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  sendFrequency?: number;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  separator?: string;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  refresh?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  height?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  minValue?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  maxValue?: number;
  /**
   *
   * @type {number}
   * @memberof WidgetDTO
   */
  step?: number;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  url?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  encoding?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  service?: string;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  period?: string;
  /**
   *
   * @type {boolean}
   * @memberof WidgetDTO
   */
  legend?: boolean;
  /**
   *
   * @type {string}
   * @memberof WidgetDTO
   */
  state?: string;
  /**
   *
   * @type {EnrichedItemDTO}
   * @memberof WidgetDTO
   */
  item?: EnrichedItemDTO;
  /**
   *
   * @type {PageDTO}
   * @memberof WidgetDTO
   */
  linkedPage?: PageDTO;
  /**
   *
   * @type {Array<WidgetDTO>}
   * @memberof WidgetDTO
   */
  widgets?: Array<WidgetDTO>;
}